pyproject.toml
[project]
name = "data-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipywidgets",
    "jupyterlab",
    "matplotlib",
    "numpy",
    "openpyxl>=3.1.5",
    "pandas",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "black",
    "deptry",
    "flake8",
    "flake8-bugbear",
    "flake8-pyproject",
    "flake8-simplify",
    "isort",
    "mypy",
    "nbqa[toolchain]",
    "pandas-stubs",
    "pre-commit",
    "pylint",
    "ruff",
    "types-beautifulsoup4",
    "types-colorama",
    "types-decorator",
    "types-jsonschema",
    "types-openpyxl",
    "types-pillow",
    "types-protobuf",
    "types-psutil",
    "types-pygments",
    "types-pyyaml",
    "types-regex",
    "types-requests",
    "types-seaborn",
    "types-setuptools",
    "types-six",
    "types-tabulate",
    "types-tqdm",
    "typing-extensions",
]

[tool.black]
line-length = 90
target-version = ["py311"]
# skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
line_length = 90
float_to_top = true
known_third_party = ["wandb"]

[tool.autoflake]
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true
recursive = true
in-place = true

[tool.flake8]
max-line-length = 90
extend-ignore = ["E203"] # Whitespace before ':' for black compatability

[tool.pylint]
py-version = "3.11"
max-line-length = 90
disable = [
  "C0116", # documentation
  "C0114", # documentation
  "C0115", # documentation
]

[tool.mypy]
python_version = "3.11"
strict = false            # it's too strict to be useful
show_error_context = true
pretty = true
# ignore_missing_imports = true # ignore packages without type-checking

[tool.ruff]
target-version = "py311"
line-length = 90

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN101",  # annotationg self in class methods, will be deprecated
  "ANN102",  # annotationg cls in class methods, will be deprecated
  "COM812",  # missing-trailing-comma
  "D",       # documentation
  "FA100",   # future-annotations
  "NPY002",  # numpy random stuff
  "PLR0913", # too many init arguments
  "RUF012",  # mutable class attributes type checking
  "T201",    # print statements
]
